requirements.txt
accelerate==1.11.0
anyio==4.11.0
av==16.0.1
bitsandbytes==0.48.2
bleach==6.3.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.0.0rc6
idna==3.11
Jinja2==3.1.6
kaggle==1.7.4.5
kagglehub==0.3.13
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.4
packaging==25.0
peft==0.18.0
pillow==12.0.0
protobuf==6.33.0
psutil==7.1.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.3
qwen-vl-utils==0.0.14
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sympy==1.14.0
text-unidecode==1.3
tokenizers==0.22.1
torch==2.9.0+cu130
torchaudio==2.9.0+cu130
torchvision==0.24.0+cu130
tqdm==4.67.1
transformers @ git+https://github.com/huggingface/transformers@64397a830188a35b5e9056036e13bb6ccfe434d8
typer-slim==0.20.0
typing_extensions==4.15.0
urllib3==2.5.0
webencodings==0.5.1
